requirements.txt
yfinance
pandas
plotly.express